refactor: consolidate describe code & remove circular dependencies
