+++ packages/salesforcedx-sobjects-faux-generator/src/describe/index.ts
@@ -5,5 +5,4 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-export { MAX_BATCH_REQUEST_SIZE, SObjectDescribe } from './sObjectDescribe';
-export { SObjectCategory, SObject, ChildRelationship, Field } from './types';
+export { SObjectDescribe } from './sObjectDescribe';
