+++ packages/salesforcedx-sobjects-faux-generator/src/generator/configUtil.ts
