+++ packages/salesforcedx-sobjects-faux-generator/src/types/index.ts
@@ -1,0 +1,16 @@
+/*
+ * Copyright (c) 2021, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+export {
+  BatchRequest,
+  BatchResponse,
+  SObject,
+  Field,
+  ChildRelationship
+} from './describe';
+
+export { SObjectCategory, SObjectRefreshSource } from './general';
